Labtest_OCR_Quaifier2_Shreya/requirements.txt
fastapi
uvicorn
pytesseract
opencv-python
python-multipart
pillow
numpy